feat: styles sentence selector
