+++ frontend/scripts/router/router.js
@@ -10,7 +10,8 @@ import {
   getPostsContent,
   getTestimonialsContent,
   getTweetsContent,
-  resetToolThunk
+  resetToolThunk,
+  loadInitialDataHome
 } from 'react-components/home/home.thunks';
 import { withSidebarNavLayout } from 'react-components/nav/sidebar-nav/with-sidebar-nav-layout.hoc';
 import { getProfileRootNodes } from 'react-components/profile-root/profile-root.thunks';
@@ -34,7 +35,12 @@ const routes = {
   home: {
     path: '/',
     page: 'home',
-    thunk: dispatchThunks(getPostsContent, getTweetsContent, getTestimonialsContent)
+    thunk: dispatchThunks(
+      getPostsContent,
+      getTweetsContent,
+      getTestimonialsContent,
+      loadInitialDataHome
+    )
   },
   explore: {
     path: '/explore',
