+++ frontend/styles/components/homepage/homepage.scss
@@ -115,6 +115,13 @@
     }
   }
 
+  .homepage-map {
+    .homepage-map-container {
+      margin: 0 auto;
+      width: 700px;
+    }
+  }
+
   .sliders {
     background-color: $white;
     padding: 60px 0;
