+++ frontend/styles/components/homepage/sentence-selector.scss
@@ -1,0 +1,43 @@
+.c-sentence-selector {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin: 60px 0 25px;
+
+  .sentence-selector-text {
+    color: $charcoal-grey;
+    font-size: 24px;
+    font-family: $font-family-2;
+    font-weight: 300;
+    line-height: 1.58;
+  }
+
+  .sentence-selector-container {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+
+  .c-dropdown {
+    margin: 0 10px;
+    min-width: 120px;
+    cursor: pointer;
+
+    .dropdown-title::after {
+      position: absolute;
+      top: 50%;
+      right: 10px;
+      border-color: $strong-pink;
+      transform: translate(-50%, calc(-50% - 1px)) rotate(45deg)
+    }
+
+    .dropdown-list {
+      top: 20px;
+      > .dropdown-item {
+        color: $charcoal-grey;
+        font-size: 13px;
+      }
+    }
+  }
+}
