+++ frontend/styles/components/shared/world-map.scss
@@ -1,5 +1,5 @@
 .c-world-map {
-  background-color: $light-gray;
+  background-color: transparent;
 
   .world-map-geography {
     fill: rgba($charcoal-grey, 0.3);
