+++ frontend/styles/homepage.scss
@@ -16,9 +16,11 @@
 'components/shared/hero',
 'components/shared/newsletter',
 'components/shared/animated-flows',
+'components/shared/world-map',
 
 'components/homepage/slider-section',
 'components/homepage/author-footer',
 'components/homepage/twitter-feed',
 'components/homepage/home-video',
+'components/homepage/sentence-selector',
 'components/homepage/homepage';
